In recovery, we prefer transition from second level state to top state, over transition from lower level state to top state(and overriding the instance preference in preference list)
In recovery from e.g. Offline to Master, we'd prefer promoting it to Slave, and at the same time promote another Slave to Master. After this is finished, switch the state of the two nodes.
This way, the system has higher availability (for Master) because it usually takes longer time to transit Offline to Slave than Slave to Master.
Therefore, we don't want to transit lower level state directly to top state, we always try to only transit second level state to top state.

For implementation, SemiAutoRebalancer and AutoRebalancer share the same logic, so remove the method in SemiAutoRebalancer, add it to AbstractRebalancer so that it’s used by both rebalancers.

After best possible state is computed, the fix compares the current state and best possible state and determines whether any pairs of  instances should switch their best possible state. See more comments in the code.

Add unit test for computeBestPossibleStateForPartition to test this feature specifically. 4 cases in total.
Modified TestRebalancePipeline so that it matches the new behavior of the rebalancer while keeping the point of the original test. Also added a new test in it specifically testing that it should not be duplicated Masters.

A couple code readability changes:
1. Simply pass in liveInstance Set, instead of ClusterDataCache, since that’s the only thing in the cache that is used, change corresponding method calls in DelayedAutoRebalancer and TestAutoRebalancerStrategy.
2. Remove useless setups, comments in BaseStageTest, re-group setupStateModel() statements.

RB=908170
G=helix-reviewers
R=ghelgess,lxia,jxue,erkim
A=jxue