Be able to stop workflow when no job is running.
Currently, to stop a workflow, the target state of the workflow is set to STOP, then when each job(as a resource in ideal state) was processed in job rebalancer, it will check whether all the jobs in the workflow is done(not in IN_PROGRESS or STOPPING) and set the workflow state to be STOP.
However, if all the jobs are already done, there’s no job in ideal state to process, so the workflow state never gets a chance to be set to STOP.

This commit adds a check in workflow rebalancer to set the state when all jobs are already done.

A test is added to test specifically this case.

RB=918081
G=helix-reviewers
A=jxue