Not add IdealState back after job finished.
In computeResourceMapping(), it’s possible that some partition are deemed to be dropped but the job finished and
IdealState is deleted altogether. In this case, dropping the partitions from IdealState will add IdealState back, which
is unnecessary.

Solution is to check whether the IdealState is already deleted before dropping partitions from it.

RB=938459
BUG=DAAS-4214
G=helix-reviewers
A=jxue