Job failure waits for task to abort.
When a job is deemed to be failed, it should first abort all running tasks before marking
the job as failed and deleting its IdealState.

To implement this, FAILING state is added to represent the state when the job is being
failed and all running tasks being aborted. Specifically, when a job should be failed, it is
marked FAILING, and abort running tasks and cancel not started tasks. After all tasks
are finished, mark the job FAILED.

Add Integration tests.

Add TargetPartitionConfigs to MockTask so that tests can specify the task config for a
specific target resource partition. In TestJobFailure, a list of specified task states is read
in from json input file and configs the corresponding task for target resource partitions.
The tasks will behave according to this config. In the end of the test, the expected task
states are compared with the actual states.

RB=946752
BUG=DAAS-4196
G=helix-reviewers
A=jxue