Testing parallel job requirements does not rely on environment.
TestTaskRebalancerParallel tests
(1) whether the job running in parallel exceed the configured max
allowed parallel job
(2) no instance is assigned to multiple jobs in the same workflow but it
requires that the workflow doesn’t finish too fast so the intermediate
state of the workflow can be captured and tested.

Remove the old test and instead, make the tasks stuck in RUNNING and
test the above two requirements explicitly.

RB=948568
BUG=DAAS-4210
G=helix-reviewers
A=jxue