Test job failure high failure threshold
When job failure threshold is equal to or higher than number of
instances, if job failure mechanism only considers the tasks currently
in CurrentState, the job will never fail. This fix takes into account
the given up tasks that no longer exist on the instances.

Add an integration test for this case.

RB=950977
BUG=DAAS-4213
G=helix-reviewers
A=jxue