Add ZkClientMonitor
ZkClientMonitor is created for each ZkClient and is used to monitor
ZkClient traffic. Each monitor is automatically tag or created with a
user-defined tag. If same tag is used, a duplicate number is used in the
MBean ObjectName.

ZkHelixManager and ZkHelixPropertyStore has their built-in tag.

Read/Write/StateChangeEvent/DataChangeEvent is monitored.
Helix related ZK paths individually monitored.

RB=985261
BUG=DAAS-4505
G=helix-reviewers
A=lxia