Support write operations to clusters/clusterId endpoints, which includes:
PUT – create a new cluster with {clusterName}, it returns 200 if the cluster already exists.
DELETE – delete this cluster.
activate -- Link this cluster to a Helix super (controller) cluster, i.e, add the cluster as a resource to the super cluster.
expand -- In the case that a set of new node is added in the cluster, use this command to balance the resources on the existing instances to new added instances.
enable – enable/resume the cluster.
disable – disable/pause the cluster.

RB=1002703
BUG=DAAS-4706
G=helix-reviewers
A=jxue