Add workflow configuration to allow or disallow assigning multiple jobs to one instance.
By default, it is not allowed that Helix assigns multiple jobs in one workflow to the same instances.
If it is set to be true, the instance may start executing multiple jobs in each workflow.

When application sets max tasks throttling for the participants, allowing overlapping assignment can maximize utilization.

Extend TestTaskRebalancerParallel to verify the change.

RB=1008734
BUG=DAAS-4768
G=helix-reviewers
A=lxia,jxue