Update MBean attributes name to fit Counter or Gauge categories.
Also change the "max" monitor item to be valid for 1 hour if not replaced.
Previously, the "max" item is reset according to a fixed time window. One issue is that if the new max value comes at the last second of this window, it only survive for 1 second.

RB=1061232
G=helix-reviewers
A=lxia,jxue