Fix WorkflowConfig.Builder.fromMap() to copy WorkflowID
The legacy fromMap() method in WorkflowConfig.builder does not copy WorkflowId from the given simple fields, this cause issue when client generate a new workflowConfig from this legacy method. In this case, the workflow config that missing id will not be read by HelixDataAccessor.

RB=1070492
BUG=HELIX-499
G=helix-reviewers
A=lxia,jjwang