[HELIX-792] Add new API for getProperty and get
These new APIs will perform different behavior as before. It will throw an exception if throwException flag set to be true and at least one of path is failed to read in the list. Previous behavior is ignoring the read failure and filled with null for unread data.

RB=1227273
BUG=HELIX-792
G=helix-reviewers
A=lxia