Change error log to info/warning when replica count cannot be read in ResourceMonitor.
Some string such as ALL_LIVEINSTANCES in the replica field cannot be processed by the monitor. And those resources are not expected to be monitoring.
Currently error log on this case drags too much attention. Lower the log level to info/warning according to the Exception type.

RB=1228840
BUG=HELIX-795
G=helix-reviewers
A=lxia