Fix topstate handoff metrics.
We've confirmed a bug in the logic that calculates topstate handoff duration.
With this issue, if the previous master instance is offline, an older handoff start time could be used to calculate the duration.
This results in huge handoff duration in the Helix metrics.
This change will fix this bug. If the previous node that holds topstate replica goes to offline, the offline time will be used as the start time.

RB=1295351
G=helix-reviewers
A=lxia,hrzhang