Verify downward load balance and fix TestPartitionMovementThrottle
TestPartitionMovementThrottle was failing after the improvement was made in IntermediateCalcStage so that downward load balance will take place while recovery balance is happening. In the process of fixing the test, it was verified by hand that downward load balance is being correctly throttled as defined by the user in StateTransitionThrottleConfig. Also an appropriate parameter adjustment was made to account for both recovery and load balance happening in the same pipeline iteration.

RB=1319672
BUG=HELIX-1025
G=helix-reviewers
A=jjwang