[TASK] Expose UserContentStore in TaskDriver
There was a user request for this feature. The intended use is to allow for aggregation work reading from temporary data written by tasks, by allowing a get() of UserContentStore at the TaskDriver level. UserContentStore is a potentially useful feature that is currently under-utilized - this will enable Gobblin and other users of Task Framework to better utilize UserContentStore.

Changelist:
1\. Add getUserContentStore() in TaskDriver
2\. Add TestUserContentStore, an integration test for this feature
3\. Add descriptive JavaDoc warning the user that get() and put() methods for UserContentStore is not thread-safe

RB=1345173
BUG=HELIX-1098
G=helix-reviewers
A=jxue