Fix 2 test bugs that fails mvn test.
TestZkReconnect has a racing issue that the expected flag maybe set earlier. Change the test script to wait until the flag is configured.
BestPossibleExternalViewVerifier has an issue that when best possible state is not able to be calculated, the verifier always return false. The expected behavior is return true if the external view does not exist.

RB=1355773
G=helix-reviewers
A=hulee