Fix the issue that resource MBean may not be cleaned up when the resource is dropped.
If a resource is not successfully created on any participant, and it is removed in this situation, the corresponding MBean maybe left over by the controller.
This fix will ensure all resource MBeans that are no longer related to any living resource to be removed.

RB=1359299
BUG=HELIX-1153
G=helix-reviewers
A=hulee