[TASK] Fix removeJob to prevent partial removal of job-related ZNodes
It was observed that when a user tried to remove a job, it would be removed partially, meaning that its IdealState got removed, but other ZNodes like config and contexts were still left over. These are essentially garbage ZNodes that the Controller will never get to clean up because the IdealState is gone.

Changelist:
1. Make removeJob() return early as soon as a ZNode failure is detected

RB=1363588
BUG=HELIX-1177
G=helix-reviewers
A=jxue