[TASK] Fix a bug in stopping workflows
There was an edge case that came to our attention lately in the stop-workflow logic. The symptom we observed was that when a workflow is cancelled via TaskDriver, its tasks would flip-flop between STOPPED and RUNNING states. This was reproducible. Upon analysis, what was happening was that in AbstractTaskDispatcher, when the current Task's state is STOPPED, sometimes the Participant would try to process a message going to COMPLETE/ERROR state, and since there is no direct transition in the Task state model going from STOPPED to COMPLETE/ERROR, it was first going to the intermediate state of RUNNING. In short, STOPPED->RUNNING was taking place, and when it hits the RUNNING state, the Controller would send it back to STOPPED since the target state for the workflow is STOP.

The fix for it was to simply add a check on the current Task state and the target state of the parent workflow before we pass the requested state assignment in AbstractTaskDispatcher. A test was added to ensure that workflows' tasks are stopped properly when stopped via TaskDriver.

Changelist:
    1. Add a check in AbstractTaskDispatcher before passing the requested state transition assignment that if the current Task state is STOPPED and the target state for the workflow is STOP, do not make the assignment and just continue (NOP)
    2. Add tests in TestStopWorkflow
    3. Ensure that quota resource is released at stop time for tasks
    4. Identify a unreachable code snippet

RB=1380507
G=helix-reviewers
A=jjwang,hrzhang