[TASK] Add logging functionality in AssignableInstanceManager
In order to debug task-related inquiries and issues, we realized that it would be very helpful if we logged there was a log recording the current quota capacity of all AssignableInstances. This is for cases where we see jobs whose tasks are not getting assigned so that we could quickly rule out the possibility of bugs in quota-based scheduling.
Changelist:
    1. Add a method that logs current quota profile in a JSON format with an option flag of only displaying when there are quota types whose capacities are full
    2. Add info logs in AssignableInstanceManager

RB=1386557
G=helix-reviewers
A=jxue