Introduce Helix ZkClient factory. And use the factory to generate new ZkClient in the critical Helix components.
The motivation of this change is sharing ZkConnection as much as possible.
DedicatedZkClient: the client that uses it's own connection.
SharedZkClient: the client that uses a shared ZkConnection with other share client.

Also defining a safer client interface HelixZkClient so as to hide the internal ZkConnection.

For the critical Helix components, the plan is:
- HelixManager (CONTROLLER, PARTICIPANT, CONTROLLER_PARTICIPANT, SPECTATOR): Dedicated ZkClient
- HelixManager (ADMINISTRATOR): Shared ZkClient
- HelixPropertyStore: Shared ZkClient
- HelixZkAccess or ZkAdmin: Shared ZkClient

ZkClient Guide Line
- DedicatedZkClient
  Isolated, no latency concern.
- SharedZkClient
  Eco friendly, could introduce more latency since multiple watchers are sequentially handled in a single connection.
  Multiple clients will use the same session. So a major difference between shared clients and dedicated clients is that, closing a shared client does not close the session automatically. Given this limitation, creating ephemeral nodes using a shared ZkClient is not supported.

RB=1388662
BUG=HELIX-1223
G=helix-reviewers
A=lxia,hrzhang