TASK: Fix a bug in WorkflowAccessor
There was a bug in WorkflowAccessor where a JSON was submitted to for PUT requests, it would created a WorkflowConfig based on the JSON but fails to create a workflow with the config. This was preventing users to create workflows via REST APIs properly.
Changelist:
1. Ensure that the configs submitted are reflected in the workflow being created

RB=1412378
G=helix-reviewers
A=jxue