TASK: Remove quotaType fields from Workflow and Job Beans
For a short while, we used quota type to denote the quota type of Task Framework resources. However, we changed the design so that we are using the workflowType and jobType fields respectively. There were places where the quotaType field was left over in the codebase, and this RB cleans it up.
Changelist:
1. Remove all quotaType fields from Bean classes
2. Add the setting of workflowType in WorkflowBean when read into Workflow.Builder

RB=1402171
G=helix-reviewers
R=jjwang,lxia,hrzhang,jxue
A=jxue