TASK2.0: Add PropertyKey APIs for new ZNode structure workflow/job paths
As part of ZNode restructuring for Task Framework, we need a convenient way to generate paths to read from and write to ZooKeeper. PropertyKey was already being used for this purpose for the most part throughout Helix, so these APIs were added in PropertyKey and PropertyPathBuilder.

Changelist:
1. Add path generation APIs for Task Framework resources in PropertyKey
2. Add a unit test for the new APIs

RB=1414259
BUG=HELIX-1137
G=helix-reviewers
A=jxue