HELIX: Fix a possible NPE in loadBalance in IntermediateStateCalcStage
In isLoadBalanceDownwardForAllReplicas() in IntermediateStateCalcStage, statePriorityMap was throwing a NPE because the partition contained a replica in ERROR state, and the map did not have an entry for it. To amend the issue, Venice added the ERROR state in the state model with a priority, and Helix added checks to prevent NPEs.
Changelist:
1. Add containsKey checks in isLoadBalanceDownwardForAllReplicas()
2. Make the Controller correctly log all partitions with ERROR state replicas
3. Add HelixDefinedStates in statePriorityList if not already added

RB=1425968
BUG=HELIX-1311
G=helix-reviewers
A=jjwang