Generalize topology domain to support graceful node swap.
There was a constraint that topology domain has to contain the instance name as the final path value. This constraint prevent configuring flexilbe topology for a dynamic cluster. In case of a node swap, the topology has to be completely re-calculated.
This change decouple the domain from instance name. So even with some nodes swapped, admin can still configure with the same topology. This will ensure a stable resource partition assignment.

RB=1476948
BUG=HELIX-1424
G=helix-reviewers
A=hrzhang