Refactor AssignableInstanceManager
Current code structure is hard to have global quota number. For optimize our pipeline computation, we need to skip the workflow computation when there is no quota left. In this case, this rb contains the refactoring for AssignableInstanceManager as central place to operate quota management.

RB=1479676
BUG=HELIX-1364
G=helix-reviewers
A=hrzhang