Fix the leader election issue when an invalid controller leader node with stale session was found.
Since HelixManager refreshes the session Id only when new session event was processed, there could be a delay before the new session Id becomes visible to the election method. In these case, an invalid leader node might be created with a stale session Id. As a result, this invalid node prevents the other controllers (including the original controller with a new session Id now) from acquiring the leadership.

This change make leader check respect the ZkNode ephemeral owner Id instead of the content. Note that in this case, the node content will still contains the stale session Id, but leader check will return the correct result. The invalid node issue will be fixed as a next step. Basically, we need to discard the stale event when processing ZK events. This is not a blocking issue here.

In addition, added a DistControllerElection test case using real helix managers.

RB=1500983
BUG=HELIX-1471
G=helix-reviewers
R=kgopalak
A=hrzhang