TASK: Make TaskAssigner honor instance constraints
Previously, ThreadCountBasedTaskAssigner was assigning to all AssignableInstances. This could potentially be problematic because some users may wish to use InstanceGroupTags, in which case we must filter out instances that do not have the appropriate tags. This RB adds a logic that helps TaskAssigner honor such constraints.

Changelist:
1. TaskAssigner only assigns to AssignableInstances contained in eligible instances
2. Add a test for this logic change

RB=1502709
BUG=HELIX-1515
G=helix-reviewers
R=jjwang,lxia,hrzhang,jxue
A=jxue