TASK: Fix double-booking of tasks upon Participant disconnect
It's been observed in production use cases (Espresso FULL-AUTO migration) that when there are transient Participant connection issues, the Controller would fail to honor maxNumberOfTasksPerInstance limit. That is to say, if the user wants only 1 task from a job (limit is set to 1), Helix must assign up to 1 task onto an instance. But upon short Participant disconnects, we saw 2 tasks in RUNNING at the same time.
The cause for this is the incorrect calculation of jobConfigLimitation in AbstractTaskDispatcher. This RB fixes this by utilizing a Map (assignedPartitions) to calculate the correct number of tasks to assign.
Changelist:
1. Modify an internal data structure (assignedPartitions)
2. Fix the logic that calculates the number of tasks to assign

RB=1521567
BUG=ESPENG-20034,HELIX-1507,HELIX-1570
G=helix-reviewers
A=jxue