TASK: Fix bug where JobDispatcher does not create UserContentStore for new jobs
It was observed that there are multiple logic paths where a new job could get scheduled: 1. scheduleJobs() 2. processJobStatusUpdateAndAssignment(). When a job is being assigned by the latter, JobDispatcher would fail to create the UserContentStore for the job, causing all subsequent read/writes to this UserContentStore fail. This is a temporary fix and further refactoring of code paths would be required in order to consolidate where new jobs get scheduled.
Changelist:
1. Add UserContentStore for jobs with null contexts

RB=1595813
BUG=HELIX-1704
G=helix-reviewers
A=jxue