HELIX: Recovery balance partitions with disabled top-state replicas
Previously, disabling of partitions or disabled instances did not affect Helix's throttling logic. This was problematic because the ability to disable was designed in in order to move partitons/replicas out of the given instance as a measure to deal with unhealthy partitions/instances. This allows, for partitions that are disabled, to go into recovery balance, and when the user has not set any throttling configs for recovery balance, these types of state transitions will go through unthrottled, avoiding downtime.
Changelist:
1. Add a check for determining rebalance type for a given partition
2. Add an integration test

RB=1600920
BUG=HELIX-1541,HELIX-1654
G=helix-reviewers
A=jjwang