TASK2.0: Job scheduling core pipeline fixes
Task Framework 2.0 had stability issues and race conditions that weren't being handled correctly. Also, integration with RuntimeJobDag had some loopholes that needed to be fixed. This diff includes such fixes and improvements that makes it really show performance gains and cuts down on redundant computation.
Changelist:
1. Race condition when a job is enqueued, only the new JobConfig is updated and not the DAG
Add a two-way selective update which ensures consistency between JobConfigs and parent DAGs
2. Moved where getNextJob() is called in scheduleJobs() in WorkflowDispatcher
This ensures that once a RuntimeJobDag is rebuilt, update for jobs happens in one pipeline run, which removes any extra delay or slowness
3. Race condition where the job you got from getNextJob is for some reason not schedulable
This is due to deleting and enqueuing a job of the same name
RuntimeJobDag has the old job name, which conflicts with the dependency in the new DAG
Fixed the test: TestTaskRebalancerStopResume so that it does not enqueue a job of the same name
4. JobRebalancer was throwing an NPE when calling processJobStatusUpdateAndAssignment()
This was sometimes making the Controller hang
Added a null check for JobConfig (job could have been deleted/purged)
5. Fix bug with isWorkflowStopped
TargetState comparison was done in the opposite way
This fixes TestRecurringJobQueue's testDeletingRecurrentQueueWithHistory
Sometimes contexts do not get deleted cleanly but this does not affect correctness
6. Add TestEnqueueJobs
7. Fix unstable TestGetLastScheduledTaskExecInfo
8. Other minor style fixes

RB=1609742
BUG=HELIX-1661,HELIX-1703,HELIX-1704
G=helix-reviewers
A=jxue