Avoid lock the cache object when require a FullRefresh.
The old synchronize control logic will prevent requiring full refresh if a refresh is in progress. This may lead to a slow callback handling.In this change, we remove the original synchronize control. The current cache update logic will be able to handle gradually refreshed data. There is no need to lock the full refresh request.

RB=1652941
BUG=HELIX-1851,gcn-29329
G=helix-reviewers
A=jxue